【TodoRepositoryのテストファイル修正】～DB接続用に～

+++ src/repositories/TodoRepository.ts
@@ -1,5 +1,6 @@
-import type { TodoInput } from "../types/type";
-import { TodoUpdatedInput } from "../types/type";
+import type { TodoInput } from "../types/TodoRequest.type";
+import type { TodoUpdatedInput } from "../types/TodoRequest.type";
+import type { Todo } from "@prisma/client";
 import { PrismaClient } from "@prisma/client";
 
 const prisma = new PrismaClient();
@@ -18,7 +19,7 @@ export class TodoRepository {
   }
 
   async save(inputData: TodoInput) {
-    const todoData = await prisma.todo.create({
+    const todoData: Todo = await prisma.todo.create({
       data: {
         title: inputData.title,
         body: inputData.body,
@@ -40,46 +41,61 @@ export class TodoRepository {
     if (count < 1 || !Number.isInteger(count)) {
       throw new Error("countは1以上の整数のみ");
     }
+
     const offset = (page - 1) * count;
     const todos = await prisma.todo.findMany({
       skip: offset,
-      take: offset + count,
+      take: count,
     });
 
     return todos;
   }
 
   async find(id: number) {
-    if (typeof id !== "number" || id < 1) {
-      throw new Error("idは必須です(1以上の数値)");
-    }
     const todoItem = await prisma.todo.findUnique({
       where: {
         id: id,
       },
     });
 
+    if (!todoItem) {
+      throw new Error("存在しないIDを指定しました。");
+    }
+
     return todoItem;
   }
 
   async update({ id, title, body }: TodoUpdatedInput) {
-    const updateTodo = await prisma.todo.update({
-      where: {
-        id: id,
-      },
-      data: { title: title, body: body },
+    const updateItem = await prisma.todo.findUnique({
+      where: { id: id },
     });
 
-    return updateTodo;
+    if (!updateItem) {
+      throw new Error("存在しないIDを指定しました。");
+    }
+
+    const updatedItem = await prisma.todo.update({
+      where: { id: id },
+      data: {
+        title: title,
+        body: body,
+      },
+    });
+
+    return updatedItem;
   }
 
   async delete(id: number) {
-    const deleteTodo = await prisma.todo.delete({
-      where: {
-        id: id,
-      },
+    const deleteItem = await prisma.todo.findUnique({
+      where: { id: id },
     });
 
-    return deleteTodo;
+    if (!deleteItem) {
+      throw new Error("存在しないIDを指定しました。");
+    }
+
+    await prisma.todo.delete({
+      where: { id: id },
+    });
   }
 }
